refactoring utils to global functions
